requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Django==3.0.2
dnspython==1.16.0
email-validator==1.0.5
ez-setup==0.9
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.2
numpy==1.18.3
Pillow==7.1.2
proglog==0.1.9
psycopg2==2.8.4
pytube3==9.6.4
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
tqdm==4.45.0
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==16.7.9
Werkzeug==1.0.1
WTForms==2.3.1
